P1.c
//The commission problem//
Question - The Commission Problem includes a salesperson in the former Arizona Territory sold rifle locks, stocks and barrels made by a gunsmith in Missouri.

Cost includes

Locks- $45

Stocks- $30

Barrels $25

The salesperson had to sell at least one complete rifle per month and production limits were such that the most the salesperson could sell in a month was 70 locks, 80 stocks and 90 barrels. After each town visit, the sales person sent a telegram to the Missouri gunsmith with the number of locks, stocks and barrels sold in the town.

At the end of the month, the salesperson sent a very short telegram showing -1 lock sold. The gunsmith then knew the sales for the month were complete and computed the salesperson's commission as follows:

On sales up to (and including) $1000= 10% On the sales up to (and includes) $1800= 15% On the sales in excess of $1800= 20%

The commission program produces a monthly sales report that gave the total number of locks, stocks and barrels sold, the salesperson's total dollar sales and finally the commission.